requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
python-jose
bcrypt
pydantic
email-validator
python-multipart
cloudinary